refactor: 파일명 변경 ReviewBody.tsx -> index.tsx
ReviewBody가 /components/review의 entry point임을 명시적으로 나타내기 위해 파일명을 ReviewBody.tsx -> index.tsx 로 수정함
